% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{adstock_weibull}
\alias{adstock_weibull}
\title{Weibull adstock functions}
\usage{
adstock_weibull(x, shape, scale, windlen = NULL, type)
}
\arguments{
\item{x}{A numeric vector.}

\item{shape, scale}{Numeric. The CDF (Cumulative Distribution Function) of Weibull has
two parameters, shape & scale, and has flexible decay rate, compared to Geometric
adstock with fixed decay rate. The shape parameter controls the shape of the decay
curve. Recommended bound is c(0.0001, 2). The larger the shape, the more S-shape. The
smaller, the more L-shape. Scale controls the inflexion point of the decay curve. We
recommend very conservative bounce of c(0, 0.1), because scale increases the adstock
half-life greatly.
The PDF (Probability Density Function) of the Weibull also shape & scale as parameter
and also has flexible decay rate as Weibull CDF. The difference is that Weibull PDF
offers lagged effect. When shape > 2, the curve peaks after x = 0 and has NULL slope at
x = 0, enabling lagged effect and sharper increase and decrease of adstock, while the
scale parameter indicates the limit of the relative position of the peak at x axis; when
1 < shape < 2, the curve peaks after x = 0 and has infinite positive slope at x = 0,
enabling lagged effect and slower increase and decrease of adstock, while scale has the
same effect as above; when shape = 1, the curve peaks at x = 0 and reduces to exponential
decay, while scale controls the inflexion point; when 0 < shape < 1, the curve peaks at
x = 0 and has increasing decay, while scale controls the inflexion point. When all
possible shapes are relevant, we recommend c(0.0001, 10) as bounds for shape; when only
strong lagged effect is of interest, we recommend c(2.0001, 10) as bound for shape. In
all cases, we recommend conservative bound of c(0, 0.1) for scale. Due to the great
flexibility of Weibull PDF, meaning more freedom in hyperparameter spaces for Nevergrad
to explore, it also requires larger iterations to converge.}

\item{windlen}{An integer value. Length of modelling window.}

\item{type}{Character. Accepts "cdf" or "pdf". CDF, or cumulative density
function of the Weibull function allows changing decay rate over time in both
C and S shape, while the peak value will always stay at the first period,
meaning no lagged effect. PDF, or the probability density function, enables
peak value occuring after the first period when shape >=1, allowing lagged
effect. Run \code{plot_adstock()} to see the difference visually.}
}
\description{
The Weibull adstock is a two-parametric adstock function that allows changing
decay rate over time, as opposed to the fixed decay rate over time as in
Geometric adstock. It has two options, the cumulative density function "cdf"
or the probability density function "pdf".
}
