% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{calibrate_mmm}
\alias{calibrate_mmm}
\title{Calibration function}
\usage{
calibrate_mmm(
  decompCollect,
  calibration_input = InputCollect$calibration_input,
  paid_media_vars = InputCollect$paid_media_vars
)
}
\arguments{
\item{decompCollect}{xxx}

\item{calibration_input}{xxx}

\item{paid_media_vars}{xxx}
}
\description{
Describe function.
}
