% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_curve_reach_freq}
\alias{df_curve_reach_freq}
\title{Robyn Dataset: Reach & frequency simulated dataset}
\format{
An object of class \code{"data.frame"}
\describe{
  \item{spend_cumulated}{cumulated spend of paid media}
  \item{response_cumulated}{cumulated reach of paid media}
  \item{freq_bucket}{Frequency bucket for cumulated reach}
}
}
\usage{
data(df_curve_reach_freq)
}
\value{
data.frame

Dataframe.
}
\description{
A simulated cumulated reach and spend dataset by frequency buckets.
The headers must be kept as
\code{c("spend_cumulated", "response_cumulated", "freq_bucket")}.
}
\examples{
data(df_curve_reach_freq)
head(df_curve_reach_freq)
}
\seealso{
Other Dataset: 
\code{\link{dt_prophet_holidays}},
\code{\link{dt_simulated_weekly}}
}
\concept{Dataset}
\keyword{datasets}
