% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_prophet_holidays}
\alias{dt_prophet_holidays}
\title{Robyn Dataset: Time series}
\format{
An object of class \code{"data.table"}
\describe{
  \item{ds}{Date}
  \item{holiday}{Daily total revenue}
  \item{country}{Television}
  \item{year}{Out of home}
}
}
\usage{
data(dt_prophet_holidays)
}
\value{
data.table

Dataframe. Contains \code{prophet}'s default holidays by country.
}
\description{
Describe the dataset. When using own holidays, please keep the
header \code{c("ds", "holiday", "country", "year")}.
}
\examples{
data(dt_prophet_holidays)
head(dt_prophet_holidays)
}
\seealso{
Other Dataset: 
\code{\link{dt_simulated_weekly}}
}
\concept{Dataset}
\keyword{datasets}
