% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_simulated_weekly}
\alias{dt_simulated_weekly}
\title{Robyn Dataset: Time series}
\format{
An object of class \code{"data.frame"}
\describe{
  \item{DATE}{Date}
  \item{revenue}{Daily total revenue}
  \item{tv_S}{Television}
  \item{ooh_S}{Out of home}
  \item{...}{...}
}
}
\usage{
data(dt_simulated_weekly)
}
\value{
data.frame

Dataframe. Contains simulated dummy dataset to test and run demo.
}
\description{
Describe the dataset. Input time series should be daily, weekly or monthly.
}
\examples{
data(dt_simulated_weekly)
head(dt_simulated_weekly)
}
\seealso{
Other Dataset: 
\code{\link{dt_prophet_holidays}}
}
\concept{Dataset}
\keyword{datasets}
