% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{get_rsq}
\alias{get_rsq}
\title{Calculate R-squared}
\usage{
get_rsq(true, predicted, p = NULL, df.int = NULL)
}
\arguments{
\item{true}{xxx}

\item{predicted}{xxx}

\item{p}{xxx}

\item{df.int}{xxx}
}
\description{
Describe function.
}
