% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{hyper_limits}
\alias{hyper_limits}
\title{Check hyperparameter limits}
\usage{
hyper_limits()
}
\value{
Dataframe. Contains upper and lower bounds for each hyperparameter.
}
\description{
Reference data.frame that shows the upper and lower bounds valid
for each hyperparameter.
}
\examples{
hyper_limits()
}
