% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{hyper_names}
\alias{hyper_names}
\title{Get hyperparameter names}
\usage{
hyper_names(adstock = InputCollect$adstock, all_media = InputCollect$all_media)
}
\arguments{
\item{adstock}{Default to InputCollect$adstock}

\item{all_media}{Default to InputCollect$all_media}
}
\description{
Describe function.
}
