% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{mic_men}
\alias{mic_men}
\title{Michaelis-Menten transformation}
\usage{
mic_men(x, Vmax, Km, reverse = FALSE)
}
\arguments{
\item{x}{xxx}

\item{Vmax}{xxx}

\item{Km}{xxx}

\item{reverse}{xxx}
}
\description{
Describe function.
}
