% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_decomp}
\alias{model_decomp}
\title{Robyn decomp function}
\usage{
model_decomp(
  coefs,
  dt_modSaturated,
  x,
  y_pred,
  i,
  dt_modRollWind,
  refreshAddedStart
)
}
\arguments{
\item{coefs}{xxx}

\item{dt_modSaturated}{xxx}

\item{x}{xxx}

\item{y_pred}{xxx}

\item{i}{xxx}

\item{dt_modRollWind}{xxx}

\item{refreshAddedStart}{xxx}
}
\description{
Describe function.
}
