% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_refit}
\alias{model_refit}
\title{Calibration function}
\usage{
model_refit(x_train, y_train, lambda, lower.limits, upper.limits)
}
\arguments{
\item{x_train}{xxx}

\item{y_train}{xxx}

\item{lambda}{xxx}

\item{lower.limits}{xxx}

\item{upper.limits}{xxx}
}
\description{
Describe function.
}
