% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{plot_adstock}
\alias{plot_adstock}
\title{Adstocking help plot}
\usage{
plot_adstock(plot = TRUE)
}
\arguments{
\item{plot}{Boolean. Do you wish to return the plot?}
}
\description{
Produce example plots for both Geometric and Weibull adstock.
}
\examples{
plot_adstock()
}
