% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_placebo}
\alias{plot_placebo}
\title{Plot Placebo Test (density + violin)}
\usage{
plot_placebo(OutputCollect)
}
\arguments{
\item{OutputCollect}{A robyn_outputs() object with `$placebo` filled}
}
\value{
patchwork of ggplots
}
\description{
Plot Placebo Test (density + violin)
}
