% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{plot_saturation}
\alias{plot_saturation}
\title{Saturation help plot}
\usage{
plot_saturation(plot = TRUE)
}
\arguments{
\item{plot}{Boolean. Do you wish to return the plot?}
}
\description{
Produce example plots for the Hill saturation curve.
}
\examples{
plot_saturation()
}
