% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{prophet_decomp}
\alias{prophet_decomp}
\title{Conduct prophet decomposition}
\usage{
prophet_decomp(
  dt_transform,
  dt_holidays,
  prophet_country,
  prophet_vars,
  prophet_signs,
  factor_vars,
  context_vars,
  paid_media_vars,
  intervalType
)
}
\value{
A list containing all prophet decomposition output.
}
\description{
When \code{prophet_vars} in \code{robyn_inputs()} is specified, this
function decomposes trend, season, holiday and weekday from the
dependent varibale.
}
