% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ridge_lambda}
\alias{ridge_lambda}
\title{Ridge lambda sequence function}
\usage{
ridge_lambda(x, y, seq_len = 100, lambda_min_ratio = 1e-04)
}
\arguments{
\item{x}{xxxxxx}

\item{y}{xxx}

\item{seq_len}{xxx}

\item{lambda_min_ratio}{xxx}
}
\description{
Describe function.
}
