% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocator.R
\name{robyn_allocator}
\alias{robyn_allocator}
\title{Budget allocator}
\usage{
robyn_allocator(
  robyn_object = NULL,
  select_run = NULL,
  InputCollect = NULL,
  OutputCollect = NULL,
  select_model = NULL,
  optim_algo = "SLSQP_AUGLAG",
  expected_spend = NULL,
  expected_spend_days = NULL,
  channel_constr_low = 0.5,
  channel_constr_up = 2,
  scenario = "max_historical_response",
  maxeval = 1e+05,
  constr_mode = "eq"
)
}
\arguments{
\item{robyn_object}{xxx}

\item{select_run}{xxx}

\item{InputCollect}{xxx}

\item{OutputCollect}{xxx}

\item{select_model}{xxx}

\item{optim_algo}{xxx}

\item{expected_spend}{xxx}

\item{expected_spend_days}{xxx}

\item{channel_constr_low}{xxx}

\item{channel_constr_up}{xxx}

\item{scenario}{xxx}

\item{maxeval}{xxx}

\item{constr_mode}{xxx}
}
\value{
List object
}
\description{
Describe function.
}
\examples{

}
