% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{robyn_clusters}
\alias{robyn_clusters}
\title{Reduce number of models based on ROI clusters and minimum combined errors}
\usage{
robyn_clusters(
  input,
  all_media = NULL,
  k = "auto",
  limit = 1,
  weights = rep(1, 3),
  dim_red = "PCA",
  quiet = FALSE,
  export = FALSE,
  ...
)
}
\arguments{
\item{input}{\code{robyn_export()}'s output or \code{pareto_aggregated.csv} results.}

\item{all_media}{A character vector. Default to \code{InputCollect$all_media}.
Includes \code{InputCollect$paid_media_vars} and \code{InputCollect$organic_vars}.}

\item{k}{Integer. Number of clusters}

\item{limit}{Integer. Top N results per cluster. If kept in "auto", will select k
as the cluster in which the WSS variance was less than 5\%.}

\item{weights}{Vector, size 3. How much should each error weight?
Order: nrmse, decomp.rssd, mape. The highest the value, the closer it will be scaled
to origin. Each value will be normalized so they all sum 1.}

\item{dim_red}{Character. Select dimensionality reduction technique.
Pass any of: \code{c("PCA", "tSNE", "all", "none")}.}

\item{quiet}{Boolean. Keep quiet? If not, print messages.}

\item{export}{Export plots into local files?}

\item{...}{Additional parameters passed to \code{lares::clusterKmeans()}.}
}
\description{
The \code{robyn_clusters()} function uses output from \code{robyn_run()},
to reduce the amount of models and help the user pick up the best (lowest
combined error) of different kinds (clusters) of models.
}
\examples{
\dontrun{
cls <- robyn_clusters(input = OutputCollect,
                      all_media = InputCollect$all_media,
                      k = 3, limit = 2,
                      weights = c(1, 1, 1.5))
}
}
\author{
Bernardo Lares (bernardolares@fb.com)
}
