% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{robyn_csv}
\alias{robyn_csv}
\title{Output results into local files: CSV files}
\usage{
robyn_csv(OutputCollect, csv_out = NULL, export = TRUE)
}
\arguments{
\item{OutputCollect}{\code{robyn_run()} outcomes.}

\item{csv_out}{Character. Accepts "pareto" or "all". Default to "pareto". Set
to "all" will output all iterations as csv. Set NULL to skip exports into CSVs.}

\item{export}{Boolean. Export results into local files.}
}
\description{
Pack \code{robyn_pareto()}, \code{robyn_plots()}, \code{robyn_csv()},
and \code{robyn_cluster()} outcomes for \code{robyn_run()} results.
}
