% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{robyn_engineering}
\alias{robyn_engineering}
\title{Transform input data}
\usage{
robyn_engineering(InputCollect = InputCollect, refresh = FALSE)
}
\arguments{
\item{InputCollect}{Default to InputCollect}

\item{refresh}{Default to FALSE. TRUE when using in robyn_refresh()}
}
\description{
Describe function.
}
