% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{robyn_json}
\alias{robyn_json}
\title{Input Data Check & Transformation}
\usage{
robyn_json(
  InputCollect,
  OutputCollect = NULL,
  select_model = NULL,
  dir = getwd()
)
}
\arguments{
\item{InputCollect}{\code{robyn_run()} outcomes.}

\item{OutputCollect}{\code{robyn_run(..., export = FALSE)} output.}

\item{select_model}{Character. Which model (by \code{solID}) do you
want to export into a JSON file?}

\item{dir}{Character. Existing directory to export JSON file to.}
}
\value{
(invisible) List. Contains all inputs and outputs of exported model.
Class: \code{robyn_json}.
}
\description{
\code{robyn_json()} generates a JSON file with all the information
required to replicate a single Robyn model.
}
\examples{
\dontrun{
InputCollectJSON <- robyn_inputs(
  dt_input = Robyn::dt_simulated_weekly,
  dt_holidays = Robyn::dt_prophet_holidays,
  json_file = "~/Desktop/RobynModel-1_29_12.json"
)
print(InputCollectJSON)
}
}
