% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{robyn_outputs}
\alias{robyn_outputs}
\title{Output results into local files}
\usage{
robyn_outputs(
  InputCollect,
  OutputModels,
  pareto_fronts = 1,
  calibration_constraint = 0.1,
  plot_folder = getwd(),
  plot_folder_sub = NULL,
  plot_pareto = TRUE,
  csv_out = "pareto",
  clusters = TRUE,
  selected = "clusters",
  ui = FALSE,
  export = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{InputCollect, OutputModels}{\code{robyn_run()} outcomes.}

\item{pareto_fronts}{Integer. Number of Pareto fronts for the output.
\code{pareto_fronts = 1} returns the best models trading off \code{NRMSE} &
\code{DECOMP.RSSD}. Increase \code{pareto_fronts} to get more model choices.}

\item{calibration_constraint}{Numeric. Default to 0.1 and allows 0.01-0.1. When
calibrating, 0.1 means top 10% calibrated models are used for pareto-optimal
selection. Lower \code{calibration_constraint} increases calibration accuracy.}

\item{plot_folder}{Character. Path for saving plots. Default
to \code{robyn_object} and saves plot in the same directory as \code{robyn_object}.}

\item{plot_folder_sub}{Character. Customize sub path to save plots. The total
path is created with \code{dir.create(file.path(plot_folder, plot_folder_sub))}.
For example, plot_folder_sub = "sub_dir".}

\item{plot_pareto}{Boolean. Set to \code{FALSE} to deactivate plotting
and saving model one-pagers. Used when testing models.}

\item{csv_out}{Character. Accepts "pareto" or "all". Default to "pareto". Set
to "all" will output all iterations as csv. Set NULL to skip exports into CSVs.}

\item{clusters}{Boolean. Apply \code{robyn_clusters()} to output models?}

\item{selected}{Character vector. Which models (by \code{solID}) do you
wish to plot the one-pagers and export? Default will take top
\code{robyn_clusters()} results.}

\item{ui}{Boolean. Save additional outputs for UI usage. List outcome.}

\item{export}{Boolean. Export outcomes into local files?}

\item{quiet}{Boolean. Keep messages off?}

\item{...}{Additional parameters passed to \code{robyn_clusters()}}
}
\value{
(Invisible) list with processed results.
}
\description{
Pack \code{robyn_plots()}, \code{robyn_csv()}, and \code{robyn_clusters()}
outcomes for \code{robyn_run()} results. When \code{UI=TRUE}, enriched
\code{OutputModels} results with additional plots and objects.
}
