% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{robyn_refresh}
\alias{robyn_refresh}
\title{Build refresh model}
\usage{
robyn_refresh(
  robyn_object,
  dt_input = dt_input,
  dt_holidays = dt_holidays,
  refresh_steps = 4,
  refresh_mode = "manual",
  refresh_iters = 100,
  refresh_trials = 1,
  plot_pareto = TRUE
)
}
\arguments{
\item{robyn_object}{xxx}

\item{dt_input}{xxx}

\item{dt_holidays}{xxx}

\item{refresh_steps}{xxx}

\item{refresh_mode}{xxx}

\item{refresh_iters}{xxx}

\item{refresh_trials}{xxx}

\item{plot_pareto}{xxx}
}
\value{
List object
}
\description{
Describe function.
}
\examples{

}
