% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{robyn_save}
\alias{robyn_save}
\title{Save Robyn object}
\usage{
robyn_save(robyn_object, select_model, InputCollect, OutputCollect)
}
\arguments{
\item{robyn_object}{xxx}

\item{select_model}{xxx}

\item{InputCollect}{xxx}

\item{OutputCollect}{xxx}
}
\value{
List object
}
\description{
Describe function.
}
\examples{

}
