% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{saturation_hill}
\alias{saturation_hill}
\title{Hill saturation function}
\usage{
saturation_hill(x, alpha, gamma, x_marginal = NULL)
}
\arguments{
\item{x}{xxx}

\item{alpha}{xxx}

\item{gamma}{xxx}

\item{x_marginal}{xxxxxx}
}
\description{
Describe function.
}
