% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{set_default_hyppar}
\alias{set_default_hyppar}
\title{Set default hyperparameters}
\usage{
set_default_hyppar(
  adstock = NULL,
  all_media = NULL,
  list_default = list(alpha = c(0.5, 3), gamma = c(0.01, 1), theta = c(0, 0.8), shape =
    c(0, 10), scale = c(0, 0.1), train_size = c(0.5, 0.9))
)
}
\arguments{
\item{adstock}{Character. InputCollect$adstock}

\item{all_media}{Character. Provide InputCollect$all_media.}

\item{list_default}{A List. Default ranges for hyperparameters.}
}
\value{
List. Expanded range of hyperparameters for all media.
}
\description{
For quick setting of hyperparameter ranges.
}
