% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{set_holidays}
\alias{set_holidays}
\title{Detect and set date variable interval}
\usage{
set_holidays(dt_transform, dt_holidays, intervalType)
}
\arguments{
\item{dt_transform}{A data.frame. Transformed input data.}

\item{dt_holidays}{A data.frame. Raw input holiday data.}

\item{intervalType}{A character. Accepts one of the values:
\code{c("day","week","month")}}
}
\value{
List. Containing the all spend-exposure model results.
}
\description{
Robyn only accepts daily, weekly and monthly data. This function
is only called in \code{robyn_engineering()}.
}
