% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transformations}
\alias{transformations}
\alias{mic_men}
\alias{run_transformations}
\title{Michaelis-Menten Transformation}
\usage{
mic_men(x, Vmax, Km, reverse = FALSE)

run_transformations(InputCollect, hyperparameters, ...)
}
\arguments{
\item{x}{Numeric value or vector. Input media spend when
\code{reverse = FALSE}. Input media exposure metrics (impression, clicks,
GRPs, etc.) when \code{reverse = TRUE}.}

\item{Vmax}{Numeric Indicates maximum rate achieved by the system.}

\item{Km}{Numeric. The Michaelis constant.}

\item{reverse}{Boolean. Input media spend when \code{reverse = FALSE}.
Input media exposure metrics (impression, clicks, GRPs etc.) when \code{reverse = TRUE}.}

\item{InputCollect}{Default to NULL. \code{robyn_inputs}'s output when
\code{hyperparameters} are not yet set.}

\item{hyperparameters}{List. Contains hyperparameter lower and upper bounds.
Names of elements in list must be identical to output of \code{hyper_names()}.
To fix hyperparameter values, provide only one value.}

\item{...}{Additional parameters passed to \code{prophet} functions.}
}
\value{
Numeric values. Transformed values.
}
\description{
The Michaelis-Menten \code{mic_men()} function is used to fit the spend
exposure relationship for paid media variables, when exposure metrics like
impressions, clicks or GRPs are provided in \code{paid_media_vars} instead
of spend metric.
}
\examples{
mic_men(x = 5:10, Vmax = 5, Km = 0.5)
}
\seealso{
Other Transformations: 
\code{\link{adstock_geometric}()},
\code{\link{saturation_hill}()}
}
\concept{Transformations}
