% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{adstock_geometric}
\alias{adstock_geometric}
\title{Geometric adstocking function}
\usage{
adstock_geometric(x, theta)
}
\arguments{
\item{x}{xxx}

\item{theta}{xxx}
}
\description{
Describe function.
}
