% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{adstock_geometric}
\alias{adstock_geometric}
\title{Geometric adstocking function}
\usage{
adstock_geometric(x, theta)
}
\arguments{
\item{x}{A numeric vector.}

\item{theta}{A numeric value. Indicates the decay rate of the geometric function.}
}
\description{
The geometric adstock function is the classic one-parametric adstock function.
}
