% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{adstock_weibull}
\alias{adstock_weibull}
\title{Weibull adstock function}
\usage{
adstock_weibull(x, shape, scale)
}
\arguments{
\item{x}{xxx}

\item{shape}{xxx}

\item{scale}{xxx}
}
\description{
Describe function.
}
