% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{adstock_weibull}
\alias{adstock_weibull}
\title{Weibull adstock function}
\usage{
adstock_weibull(x, shape, scale)
}
\arguments{
\item{x}{A numeric vector.}

\item{shape}{A numeric value. The shape parameter controls the decay shape.
The larger the shape value, the more S-shape the curve shows. The smaller the
shape value, the more L-shape the curve shows.}

\item{scale}{A numeric value. The scale parameter controls the decay inflexion
point. The larger the scale value, the later the inflexion point occurs.}
}
\description{
The Weibull adstock function is a two-parametric adstock function that allows
changing decay rate over time, compared to the fixed decay rate over time as
in Geomeric adstock.
}
