% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_prophet_holidays}
\alias{dt_prophet_holidays}
\title{Robyn Dataset: Holidays by Country}
\format{
An object of class \code{"data.frame"}
\describe{
  \item{ds}{Date}
  \item{holiday}{Name of celebrated holiday}
  \item{country}{Code for the country (Alpha-2)}
  \item{year}{Year of \code{ds}}
}
}
\usage{
data(dt_prophet_holidays)
}
\value{
data.frame

Dataframe. Contains \code{prophet}'s default holidays by country.
}
\description{
Contains \code{prophet}'s "new" default holidays by country.
When using own holidays, please keep the header
\code{c("ds", "holiday", "country", "year")}.
}
\examples{
data(dt_prophet_holidays)
head(dt_prophet_holidays)
}
\seealso{
Other Dataset: 
\code{\link{dt_simulated_weekly}}
}
\concept{Dataset}
\keyword{datasets}
