% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{hyper_names}
\alias{hyper_names}
\title{Get correct hyperparameter names}
\usage{
hyper_names(adstock, all_media)
}
\arguments{
\item{adstock}{A character. Default to \code{InputCollect$adstock}.
Accepts "geometric" or "weibull"}

\item{all_media}{A character vector. Default to \code{InputCollect$all_media}.
Includes \code{InputCollect$paid_media_vars} and \code{InputCollect$organic_vars}.}
}
\description{
Output all hyperparameter names and help specifying the list of hyperparameters that
is inserted into the \code{hyperparameters = } parameter in \code{robyn_inputs()}
}
\details{
### Guide to setup hyperparameters

## 1. get correct hyperparameter names:
All variables in \code{paid_media_vars} or \code{organic_vars} require hyperprameters
and will be transformed by adstock & saturation. Difference between \code{organic_vars}
and \code{organic_vars} is that \code{paid_media_vars} has spend that
needs to be specified in \code{paid_media_spends} specifically. Run \code{hyper_names()}
to get correct hyperparameter names. All names in hyperparameters must
equal names from \code{hyper_names()}, case sensitive.

## 2. get guidance for setting hyperparameter bounds:
For geometric adstock, use theta, alpha & gamma. For weibull adstock,
use shape, scale, alpha, gamma.

Theta: In geometric adstock, theta is decay rate. guideline for usual media genre:
TV c(0.3, 0.8), OOH/Print/Radio c(0.1, 0.4), digital c(0, 0.3)

Shape: In weibull adstock, shape controls the decay shape. Recommended c(0.0001, 2).
The larger, the more S-shape. The smaller, the more L-shape. Channel-type specific
values still to be investigated

Scale: In weibull adstock, scale controls the decay inflexion point. Very conservative
recommended bounce c(0, 0.1), becausee scale can increase adstocking half-life greaetly.
Channel-type specific values still to be investigated

Alpha: In s-curve transformation with hill function, alpha controls the shape between
exponential and s-shape. Recommended c(0.5, 3). The larger the alpha, the more S-shape.
The smaller, the more C-shape

Gamma: In s-curve transformation with hill function, gamma controls the inflexion point.
Recommended bounce c(0.3, 1). The larger the gamma, the later the inflection point
in the response curve

Helper plots:
Run \code{plot_adstock(TRUE)} to get adstock transformation example plot,
helping you understand geometric/theta and weibull/shape/scale transformation
Run \code{plot_saturation(TRUE)} to get saturation curve transformation example plot,
helping you understand hill/alpha/gamma transformation

## 3. set each hyperparameter bounds. They either contains two values e.g. c(0, 0.5),
or only one value (in which case you've "fixed" that hyperparameter)
}
\examples{
hyper_names("geometric", all_media = LETTERS[1:3])
hyper_names("weibull", all_media = LETTERS[1:2])
}
