% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{hyper_names}
\alias{hyper_names}
\title{Get hyperparameter names}
\usage{
hyper_names(adstock, all_media)
}
\arguments{
\item{adstock}{Default to \code{InputCollect$adstock}}

\item{all_media}{Default to \code{InputCollect$all_media}}
}
\description{
Describe function.
}
