% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{mic_men}
\alias{mic_men}
\title{Michaelis-Menten Transformation}
\usage{
mic_men(x, Vmax, Km, reverse = FALSE)
}
\arguments{
\item{x}{Numeric value or vector. Input media spend when
\code{reverse = FALSE}. Input media exposure metrics (impression, clicks,
GRPs, etc.) when \code{reverse = TRUE}.}

\item{Vmax}{Numeric Indicates maximum rate achieved by the system.}

\item{Km}{Numeric. The Michaelis constant.}

\item{reverse}{Boolean. Input media spend when \code{reverse = FALSE}.
Input media exposure metrics (impression, clicks, GRPs etc.) when \code{reverse = TRUE}.}
}
\description{
The Michaelis-Menten \code{mic_men()} function is used to fit the spend
exposure relationship for paid media variables, when exposure metrics like
impressions, clicks or GRPs are provided in \code{paid_media_vars} instead
of spend metric.
}
\seealso{
Other Transformations: 
\code{\link{adstock_geometric}()},
\code{\link{saturation_hill}()}
}
\concept{Transformations}
