% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{plot_adstock}
\alias{plot_adstock}
\title{Adstocking help plot}
\usage{
plot_adstock(plotAdstockCurves)
}
\arguments{
\item{plotAdstockCurves}{A logical value.}
}
\description{
Produce example plots for both Geometric and Weibull adstock.
}
