% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{Robyn}
\alias{Robyn}
\alias{Robyn-package}
\title{Robyn MMM Project from Meta Marketing Science}
\description{
Robyn is an automated Marketing Mix Modeling (MMM) code. It aims to reduce human
bias by means of ridge regression and evolutionary algorithms, enables actionable
decision making providing a budget allocator and diminishing returns curves and
allows ground-truth calibration to account for causation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/facebookexperimental/Robyn}
  \item \url{https://facebookexperimental.github.io/Robyn/}
  \item Report bugs at \url{https://github.com/facebookexperimental/Robyn/issues}
}

}
\author{
Gufeng Zhou (gufeng@meta.com)

Leonel Sentana (leonelsentana@meta.com)

Igor Skokan (igorskokan@meta.com)

Bernardo Lares (bernardolares@meta.com)
}
