% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{robyn_converge}
\alias{robyn_converge}
\title{Check model convergence}
\usage{
robyn_converge(OutputModels, n_cuts = 10, threshold_sd = 0.025)
}
\arguments{
\item{OutputModels}{List. Output from \code{robyn_run()}}

\item{n_cuts}{Integer. Default to 10. Convergence is calculated on last
quantile of cuts.}

\item{threshold_sd}{Numeric. Default to 0.025 that is empirically derived.}
}
\description{
The \code{robyn_converge()} function consumes output from \code{robyn_run()}
and calculate convergence status and output convergence plots.
}
\examples{
\dontrun{
OutputModels <- robyn_converge(
  OutputModels = OutputModels,
  n_cuts = 10,
  threshold_sd = 0.025
)
}
}
