% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{robyn_csv}
\alias{robyn_csv}
\title{Output results into local files: CSV files}
\usage{
robyn_csv(OutputCollect, csv_out = NULL, export = TRUE)
}
\arguments{
\item{OutputCollect}{\code{robyn_run(..., export = FALSE)} output}

\item{csv_out}{Character. Accepts "pareto" or "all". Default to "pareto". Set
to "all" will output all iterations as csv. Set NULL to skip exports into CSVs.}

\item{export}{Boolean. Export outcomes into local files?}
}
\description{
Output results into local files: CSV files
}
