% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{robyn_mmm}
\alias{robyn_mmm}
\title{Major Robyn modelling function}
\usage{
robyn_mmm(
  hyper_collect,
  InputCollect,
  iterations = InputCollect$iterations,
  lambda.n = 100,
  lambda_fixed = NULL,
  refresh = FALSE
)
}
\arguments{
\item{hyper_collect}{List object}

\item{InputCollect}{xxx}

\item{iterations}{xxx}

\item{lambda.n}{xxx}

\item{lambda_fixed}{xxx}

\item{refresh}{xxx}
}
\description{
Describe function.
}
\examples{
xxx
}
