% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{robyn_mmm}
\alias{robyn_mmm}
\title{The core MMM function}
\usage{
robyn_mmm(
  hyper_collect,
  InputCollect,
  iterations = InputCollect$iterations,
  lambda.n = 100,
  lambda_fixed = NULL,
  refresh = FALSE
)
}
\arguments{
\item{hyper_collect}{A list containing hyperparameter bounds. Defaults to
\code{InputCollect$hyperparameters}.}

\item{InputCollect}{A List containing all input parameters for the model.}

\item{iterations}{A integer. Number of iterations to run.}

\item{lambda.n}{A integer. Number of lambda cross-validation in glmnet.
Defaults to 100.}

\item{lambda_fixed}{A logical. \code{lambda_fixed = TRUE} when inputting
old model results}

\item{refresh}{A logical value. Set to \code{TRUE} when used in \code{robyn_refresh()}}
}
\description{
The \code{robyn_mmm()} function activates Nevergrad to generate samples of
hyperparameters, conducts media transformation within each loop, fits the
Ridge regression, calibrates the model optionally, decomposes responses
and collects the result. It's an inner function within \code{robyn_run()}.
}
