% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{robyn_plots}
\alias{robyn_plots}
\title{Generate and Export Robyn Plots}
\usage{
robyn_plots(InputCollect, OutputCollect, export = TRUE)
}
\arguments{
\item{InputCollect}{\code{robyn_run()} outcomes.}

\item{OutputCollect}{\code{robyn_run(..., export = FALSE)} output}

\item{export}{Boolean. Export outcomes into local files?}
}
\description{
Generate and Export Robyn Plots
}
