% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{robyn_response}
\alias{robyn_response}
\title{Response Function}
\usage{
robyn_response(
  InputCollect = NULL,
  OutputCollect = NULL,
  json_file = NULL,
  robyn_object = NULL,
  select_build = NULL,
  media_metric = NULL,
  select_model = NULL,
  metric_value = NULL,
  dt_hyppar = NULL,
  dt_coef = NULL,
  quiet = FALSE,
  ...
)
}
\description{
Response Function
}
