% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{robyn_response}
\alias{robyn_response}
\title{The response function}
\usage{
robyn_response(
  robyn_object = NULL,
  select_run = NULL,
  paid_media_var = NULL,
  select_model = NULL,
  Spend = NULL,
  dt_hyppar = NULL,
  dt_coef = NULL,
  InputCollect = NULL
)
}
\arguments{
\item{robyn_object}{xxx}

\item{select_run}{xxx}

\item{paid_media_var}{xxx}

\item{select_model}{xxx}

\item{Spend}{xxx}

\item{dt_hyppar}{xxx}

\item{dt_coef}{xxx}

\item{InputCollect}{xxxxxx}
}
\description{
The \code{robyn_mmm()} function activates Nevergrad to generate samples of
hyperparameters, conducts media transformation within each loop, fits the
Ridge regression, calibrates the model optionally, decomposes responses
and collects the result. It's an inner function within \code{robyn_run()}.
}
\examples{

}
