% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{robyn_run}
\alias{robyn_run}
\title{Major Robyn modelling function}
\usage{
robyn_run(
  InputCollect,
  plot_folder = getwd(),
  dt_hyper_fixed = NULL,
  pareto_fronts = 1,
  plot_pareto = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{InputCollect}{List object}

\item{plot_folder}{xxx}

\item{dt_hyper_fixed}{xxx}

\item{pareto_fronts}{xxx}

\item{plot_pareto}{xxx}

\item{refresh}{xxx}
}
\description{
Describe function.
}
\examples{
xxx
}
