% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{robyn_run}
\alias{robyn_run}
\title{The major Robyn modelling function}
\usage{
robyn_run(
  InputCollect,
  plot_folder = getwd(),
  pareto_fronts = 1,
  plot_pareto = TRUE,
  refresh = FALSE,
  dt_hyper_fixed = NULL,
  ui = FALSE
)
}
\arguments{
\item{InputCollect}{List. Contains all input parameters for the model.
Required when \code{robyn_object} is not provided.}

\item{plot_folder}{Character. Path for saving plots. Default
to \code{robyn_object} and saves plot in the same directory as \code{robyn_object}.}

\item{pareto_fronts}{Integer. Number of Pareto fronts for the output.
\code{pareto_fronts = 1} returns the best models trading off \code{NRMSE} &
\code{DECOMP.RSSD}. Increase \code{pareto_fronts} to get more model choices.}

\item{plot_pareto}{Boolean. Set to \code{FALSE} to deactivate plotting
and saving model one-pagers. Used when testing models.}

\item{refresh}{Boolean. Set to \code{TRUE} when used in \code{robyn_refresh()}}

\item{dt_hyper_fixed}{data.frame. Only provide when loading old model results.
It consumes hyperparameters from saved csv \code{pareto_hyperparameters.csv}.}

\item{ui}{Boolean. Save additional outputs for UI usage. List outcome.}
}
\description{
The \code{robyn_run()} function consumes output from \code{robyn_input()},
runs the \code{robyn_mmm()} functions and plots and collects the result.
}
\examples{
\dontrun{
OutputCollect <- robyn_run(
  InputCollect = InputCollect,
  plot_folder = robyn_object,
  pareto_fronts = 3,
  plot_pareto = TRUE
)
}
}
