% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R, R/refresh.R
\name{robyn_load}
\alias{robyn_load}
\alias{robyn_save}
\alias{print.robyn_save}
\alias{plot.robyn_save}
\title{Export Robyn Model to Local File}
\usage{
robyn_load(robyn_object, select_build = NULL, quiet = FALSE)

robyn_save(
  robyn_object,
  select_model,
  InputCollect,
  OutputCollect,
  quiet = FALSE
)

\method{print}{robyn_save}(x, ...)

\method{plot}{robyn_save}(x, ...)
}
\arguments{
\item{robyn_object}{Character or List. Path of the \code{Robyn.RDS} object
that contains all previous modeling information or the imported list.}

\item{select_build}{Integer. Default to the latest model build. \code{select_build = 0}
selects the initial model. \code{select_build = 1} selects the first refresh model.}

\item{quiet}{Boolean. Keep messages off?}

\item{select_model}{Character. A model \code{SolID}. When \code{robyn_object}
is provided, \code{select_model} defaults to the already selected \code{SolID}. When
\code{robyn_object} is not provided, \code{select_model} must be provided with
\code{InputCollect} and \code{OutputCollect}, and must be one of
\code{OutputCollect$allSolutions}.}

\item{InputCollect}{List. Contains all input parameters for the model.
Required when \code{robyn_object} is not provided.}

\item{OutputCollect}{List. Containing all model result.
Required when \code{robyn_object} is not provided.}

\item{x}{\code{robyn_save()} output.}

\item{...}{Additional parameters passed to \code{robyn_outputs()}.}
}
\value{
(Invisible) list with imported results

(Invisible) list with filename and summary. Class: \code{robyn_save}.
}
\description{
Use \code{robyn_save()} to select and save as .RDS file the initial model.
}
