% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{robyn_save}
\alias{robyn_save}
\title{Export Robyn Model to Local File}
\usage{
robyn_save(robyn_object, InputCollect, OutputCollect, select_model = NULL)
}
\arguments{
\item{robyn_object}{Character. Path of the \code{Robyn.RDS} object
that contains all previous modeling information.}

\item{InputCollect}{List. Contains all input parameters for the model.
Required when \code{robyn_object} is not provided.}

\item{OutputCollect}{List. Containing all model result.
Required when \code{robyn_object} is not provided.}

\item{select_model}{Character. A model \code{SolID}. When \code{robyn_object}
is provided, \code{select_model} defaults to the already selected \code{SolID}. When
\code{robyn_object} is not provided, \code{select_model} must be provided with
\code{InputCollect} and \code{OutputCollect}, and must be one of
\code{OutputCollect$allSolutions}.}
}
\value{
(Invisible) file's name.
}
\description{
Use \code{robyn_save()} to select and save as .RDS file the initial model.
}
\examples{
\dontrun{
# Get model IDs from OutputCollect
select_model <- "3_10_3"

# Save the results. Overwriting old object needs confirmation.
robyn_save(
  robyn_object = "~/Desktop/Robyn.RDS",
  InputCollect = InputCollect,
  OutputCollect = OutputCollect,
  select_model = select_model
)
}
}
