% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{robyn_save}
\alias{robyn_save}
\title{Save Robyn object}
\usage{
robyn_save(robyn_object, select_model, InputCollect, OutputCollect)
}
\arguments{
\item{robyn_object}{Character. Path of the \code{Robyn.RDS} object
that contains all previous modeling information.}

\item{select_model}{Character. A model \code{SolID}. When \code{robyn_object}
is provided, \code{select_model} defaults to the already selected \code{SolID}. When
\code{robyn_object} is not provided, \code{select_model} must be provided with
\code{InputCollect} and \code{OutputCollect}, and must be one of
\code{OutputCollect$allSolutions}.}

\item{InputCollect}{List. Contains all input parameters for the model.
Required when \code{robyn_object} is not provided.}

\item{OutputCollect}{List. Containing all model result.
Required when \code{robyn_object} is not provided.}
}
\value{
A list containing all information for the initial model.
}
\description{
Use \code{robyn_save()} to select and save the initial model.
}
\examples{
\dontrun{
## Get all model IDs in result from OutputCollect$allSolutions

## Select one from above
select_model <- "3_10_3"

## Save the robyn object. Overwriting old object needs confirmation.
robyn_object <- "~/Desktop/Robyn.RDS"
robyn_save(
  robyn_object = robyn_object,
  select_model = select_model,
  InputCollect = InputCollect,
  OutputCollect = OutputCollect
)
}
}
