% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{robyn_save}
\alias{robyn_save}
\alias{print.robyn_save}
\alias{plot.robyn_save}
\alias{robyn_load}
\title{Export Robyn Model to Local File [DEPRECATED]}
\usage{
robyn_save(
  InputCollect,
  OutputCollect,
  robyn_object = NULL,
  select_model = NULL,
  quiet = FALSE
)

\method{print}{robyn_save}(x, ...)

\method{plot}{robyn_save}(x, ...)

robyn_load(robyn_object, select_build = NULL, quiet = FALSE)
}
\arguments{
\item{InputCollect}{List. Contains all input parameters for the model.
Required when \code{robyn_object} is not provided.}

\item{OutputCollect}{List. Containing all model result.
Required when \code{robyn_object} is not provided.}

\item{robyn_object}{Character or List. Path of the \code{Robyn.RDS} object
that contains all previous modeling information or the imported list.}

\item{select_model}{Character. A model \code{SolID}. When \code{robyn_object}
is provided, \code{select_model} defaults to the already selected \code{SolID}. When
\code{robyn_object} is not provided, \code{select_model} must be provided with
\code{InputCollect} and \code{OutputCollect}, and must be one of
\code{OutputCollect$allSolutions}.}

\item{quiet}{Boolean. Keep messages off?}

\item{x}{\code{robyn_save()} output.}

\item{...}{Additional parameters passed to \code{robyn_outputs()}.}

\item{select_build}{Integer. Default to the latest model build. \code{select_build = 0}
selects the initial model. \code{select_build = 1} selects the first refresh model.}
}
\value{
(Invisible) list with filename and summary. Class: \code{robyn_save}.

(Invisible) list with imported results
}
\description{
Use \code{robyn_save()} to select and save as .RDS file the initial model.
}
