% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{robyn_train}
\alias{robyn_train}
\title{Train Robyn Models}
\usage{
robyn_train(
  InputCollect,
  dt_hyper_fixed = NULL,
  lambda_control = 1,
  refresh = FALSE,
  seed = 123,
  quiet = FALSE
)
}
\arguments{
\item{InputCollect}{List. Contains all input parameters for the model.
Required when \code{robyn_object} is not provided.}

\item{dt_hyper_fixed}{data.frame. Only provide when loading old model results.
It consumes hyperparameters from saved csv \code{pareto_hyperparameters.csv}.}

\item{lambda_control}{Numeric. From 0-1. Tunes ridge lambda between
lambda.min and lambda.1se.}

\item{refresh}{Boolean. Set to \code{TRUE} when used in \code{robyn_refresh()}.}

\item{seed}{Integer. For reproducible results when running nevergrad.}

\item{quiet}{Boolean. Keep messages off?}
}
\description{
The \code{robyn_train()} function consumes output from \code{robyn_input()}
and runs the \code{robyn_mmm()} on each trial.
}
\examples{
\dontrun{
OutputCollect <- robyn_train(
  InputCollect = InputCollect,
  dt_hyper_fixed = NULL,
  seed = 0
)
}
}
