% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{saturation_hill}
\alias{saturation_hill}
\title{Hill saturation function}
\usage{
saturation_hill(x, alpha, gamma, x_marginal = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{alpha}{A numeric value. Alpha controls the shape of the saturation curve.
The larger the alpha, the more S-shape. The smaller, the more C-shape.}

\item{gamma}{A numeric value. Gamma controls the inflexion point of the
saturation curve. The larger the gamma, the later the inflexion point occurs.}

\item{x_marginal}{A numeric value. When provided, the function returns the
Hill-transformed value of the x_marginal input.}
}
\description{
The Hill function applied here is a two-parametric version of the
function that allows the saturation curve to flip between S and C shape.
}
